+++ antpack/__init__.py
@@ -1,6 +1,8 @@
 #Used for determining the version if running setup.
 #Only change if building a new version.
-__version__ = "0.0.2"
+__version__ = "0.1.0"
 
 from .consensus_update_tools import build_consensus_alignment
 from .single_chain_annotator import SingleChainAnnotator
+from .sequence_scoring_tool import SequenceScoringTool
+from .utilities.data_compile_utils import get_position_dict