+++ antpack/constants/catmix_constants.py
@@ -1,0 +1,4 @@
+"""Stores constants used by the categorical mixture."""
+
+#The smallest allowed probability value (for numerical stability).
+MINIMUM_PROB_VAL = 1e-14