+++ redo.sh
@@ -1,0 +1,4 @@
+pip uninstall AntPack
+rm -rf build antpack.egg-info
+rm *.so
+python setup.py develop