+++ HISTORY.md
@@ -3,3 +3,8 @@ Initial release.
 
 ### Version 0.0.2
 Added Kabat, Martin schemes.
+
+### Version 0.1.0
+Added sequence scoring functionality for humanness.
+Updated all numbering tools to improve behavior in
+the presence of N-terminal deletions.